pyproject.toml
[project]
name = "sayclip"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "copykitten>=2.0.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "pywebview[pyside6]>=6.1",
]
